refactor: give runtime process liveness its own module
